[go-gen] Rename service->server and create service dir for generators etc.
