+++ src/main/resources/go/genFiles/service/util/auth_struct.go.snippet
