+++ src/main/resources/go/genFiles/service/util/ids_from_request.go.snippet
