+++ src/main/resources/go/genFiles/service/util/util.go.snippet
