+++ src/main/scala/temple/generate/server/go/service/GoServiceCommGenerator.scala
@@ -1,7 +1,6 @@
-package temple.generate.service.go
+package temple.generate.server.go.service
 
 import temple.generate.utils.CodeTerm.{CodeWrap, mkCode}
-import temple.utils.FileUtils
 
 object GoServiceCommGenerator {
