+++ src/test/scala/temple/generate/server/go/GoAuthServiceGeneratorTest.scala
@@ -1,4 +1,4 @@
-package temple.generate.service.go.auth
+package temple.generate.server.go.auth
 
 import org.scalatest.{FlatSpec, Matchers}
