+++ src/test/scala/temple/generate/server/go/testfiles/auth/auth.go.snippet
