+++ src/test/scala/temple/generate/server/go/testfiles/auth/comm/handler.go.snippet
