+++ src/test/scala/temple/generate/server/go/testfiles/auth/go.mod.snippet
