+++ src/test/scala/temple/generate/server/go/testfiles/auth/util/util.go.snippet
