+++ src/test/scala/temple/generate/server/go/testfiles/match/dao/dao.go.snippet
