+++ src/test/scala/temple/generate/server/go/testfiles/match/go.mod.snippet
