+++ src/test/scala/temple/generate/server/go/testfiles/match/match.go.snippet
