+++ src/test/scala/temple/generate/server/go/testfiles/user/dao/dao.go.snippet
