+++ src/test/scala/temple/generate/server/go/testfiles/user/dao/errors.go.snippet
