+++ src/test/scala/temple/generate/server/go/testfiles/user/go.mod.snippet
