+++ src/test/scala/temple/generate/server/go/testfiles/user/user.go.snippet
