+++ src/test/scala/temple/generate/server/go/testfiles/user/util/util.go.snippet
